+++ instrumentation/opentelemetry-instrumentation-system-metrics/tests/test_system_metrics.py
@@ -18,12 +18,13 @@ from collections import namedtuple
 from platform import python_implementation
 from unittest import mock
 
-from opentelemetry.instrumentation.system_metrics import (
-    SystemMetricsInstrumentor,
-)
 from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.metrics.export import InMemoryMetricReader
 from opentelemetry.test.test_base import TestBase
+
+from opentelemetry.instrumentation.system_metrics import (
+    SystemMetricsInstrumentor,
+)
 
 
 def _mock_netconnection():
@@ -96,7 +97,7 @@ class TestSystemMetrics(TestBase):
             for scope_metrics in resource_metrics.scope_metrics:
                 for metric in scope_metrics.metrics:
                     metric_names.append(metric.name)
-        self.assertEqual(len(metric_names), 18)
+        self.assertEqual(len(metric_names), 21)
 
         observer_names = [
             "system.cpu.time",
@@ -117,6 +118,9 @@ class TestSystemMetrics(TestBase):
             f"process.runtime.{self.implementation}.memory",
             f"process.runtime.{self.implementation}.cpu_time",
             f"process.runtime.{self.implementation}.gc_count",
+            f"process.runtime.{self.implementation}.thread_count",
+            f"process.runtime.{self.implementation}.context_switches",
+            f"process.runtime.{self.implementation}.cpu.utilization",
         ]
 
         for observer in metric_names:
@@ -128,6 +132,9 @@ class TestSystemMetrics(TestBase):
             "process.runtime.memory": ["rss", "vms"],
             "process.runtime.cpu.time": ["user", "system"],
             "process.runtime.gc_count": None,
+            "process.runtime.thread_count": None,
+            "process.runtime.cpu.utilization": None,
+            "process.runtime.context_switches": ["involuntary", "voluntary"],
         }
 
         reader = InMemoryMetricReader()
@@ -140,12 +147,15 @@ class TestSystemMetrics(TestBase):
             for scope_metrics in resource_metrics.scope_metrics:
                 for metric in scope_metrics.metrics:
                     metric_names.append(metric.name)
-        self.assertEqual(len(metric_names), 3)
+        self.assertEqual(len(metric_names), 6)
 
         observer_names = [
             f"process.runtime.{self.implementation}.memory",
             f"process.runtime.{self.implementation}.cpu_time",
             f"process.runtime.{self.implementation}.gc_count",
+            f"process.runtime.{self.implementation}.thread_count",
+            f"process.runtime.{self.implementation}.context_switches",
+            f"process.runtime.{self.implementation}.cpu.utilization",
         ]
 
         for observer in metric_names:
@@ -161,9 +171,9 @@ class TestSystemMetrics(TestBase):
                     for data_point in metric.data.data_points:
                         for expect in expected:
                             if (
-                                dict(data_point.attributes)
-                                == expect.attributes
-                                and metric.name == observer_name
+                                    dict(data_point.attributes)
+                                    == expect.attributes
+                                    and metric.name == observer_name
                             ):
                                 self.assertEqual(
                                     data_point.value,
@@ -781,4 +791,38 @@ class TestSystemMetrics(TestBase):
         ]
         self._test_metrics(
             f"process.runtime.{self.implementation}.gc_count", expected
+        )
+
+    @mock.patch("psutil.Process.num_ctx_switches")
+    def test_runtime_context_switches(self, mock_process_num_ctx_switches):
+        PCtxSwitches = namedtuple("PCtxSwitches", ["voluntary", "involuntary"])
+
+        mock_process_num_ctx_switches.configure_mock(
+            **{"return_value": PCtxSwitches(voluntary=1, involuntary=2)}
+        )
+
+        expected = [
+            _SystemMetricsResult({"type": "voluntary"}, 1),
+            _SystemMetricsResult({"type": "involuntary"}, 2),
+        ]
+        self._test_metrics(
+            f"process.runtime.{self.implementation}.context_switches", expected
+        )
+
+    @mock.patch("psutil.Process.num_threads")
+    def test_runtime_thread_num(self, mock_process_thread_num):
+        mock_process_thread_num.configure_mock(**{"return_value": 42})
+
+        expected = [_SystemMetricsResult({}, 42)]
+        self._test_metrics(
+            f"process.runtime.{self.implementation}.thread_count", expected
+        )
+
+    @mock.patch("psutil.Process.cpu_percent")
+    def test_runtime_cpu_percent(self, mock_process_cpu_percent):
+        mock_process_cpu_percent.configure_mock(**{"return_value": 42})
+
+        expected = [_SystemMetricsResult({}, 42)]
+        self._test_metrics(
+            f"process.runtime.{self.implementation}.cpu.utilization", expected
         )