+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/bootstrap_gen.py
@@ -24,6 +24,10 @@ libraries = {
         "library": "aiohttp ~= 3.0",
         "instrumentation": "opentelemetry-instrumentation-aiohttp-client==0.42b0.dev",
     },
+    "aiohttp": {
+        "library": "aiohttp ~= 3.0",
+        "instrumentation": "opentelemetry-instrumentation-aiohttp-server==0.42b0.dev",
+    },
     "aiopg": {
         "library": "aiopg >= 0.13.0, < 2.0.0",
         "instrumentation": "opentelemetry-instrumentation-aiopg==0.42b0.dev",
@@ -82,6 +86,10 @@ libraries = {
     },
     "flask": {
         "library": "flask >= 1.0, < 3.0",
+        "instrumentation": "opentelemetry-instrumentation-flask==0.42b0.dev",
+    },
+    "werkzeug": {
+        "library": "werkzeug < 3.0.0",
         "instrumentation": "opentelemetry-instrumentation-flask==0.42b0.dev",
     },
     "grpcio": {